requirements.txt
glfw
PyOpenGL
numpy
PyGLM
PyWavefront